feat(refactor): example
This is an example of a refactoring for the authentication middleware and
the authentication endpoint.

This uses the best practices and recommmandation of the golang code review.
This allows cleaner dependencies injection and testing on every level.

This PR is used to make a proposal for the sql queries to simplify our use of
pgx. I recommand using squirrel when using dynamic filters and fields and using
raw queries when possible.