Add a fallthrough option for workflow rules.
Before this commit, workflows would only perform the actions listed under the first matching rule.

While this is still the default, a new option on a rule, `fallthrough`, allows defining that workflow evaluation should continue even if this rules matches.
All the actions of of the matching rules up until the first failing or not-fallthrough rule will be executed, in order.

As we can only add a decision to a case once, only the first action of this kind will be executed, the other ones will be noops.