Fix audit logs mixing user ID and API key ID.
A race was present in the code setting the variables for user
and API key ID in the SQL session. We did not clear the previous
values.

The blast radius should be quite limited, since we do not expose writeable,
public API endpoints yet, but we should check the data.