Fix improper validation of which decisions can be added to a case.
When adding decisions to a case, we improperly validated that the
current actor has access to the decisions. This made it so someone
was able to add decisions from foreign organizations to one of
their cases.

This could not give access to the data, which was properly guarded,
but could potentially remove a decision from someone else's case,
some kind of denial of service.