Allow changing the email attribute to another claim.
This adds a configuration parameter to instruct us to look for the user's email
address from a different claim as `email`. When this is set, we will not check
for the `email_verified` claim either.

This is useful for identity providers that either do not expose the email, or
do not guarantee it is verified.